restore_config files
